[modular tests] Enable component_context test.
I ran it a few dozen times without flake, so I try to enable
it again.

Also, remove some weak pointer protection that is not needed.
Weak poiner protection is necessary only around callbacks that
are *not* cancelled when the component instance is destroyed,
i.e. tasks that are run directly on the run loop. Everything
that is a callback of a fidl method on a fidl interface ptr
that is a data member of |this| does not need weak pointer
protection.

TESTED=run_modular_tests.sh component_context 50x on device; SQ

Change-Id: Id3e7f32f6c1e9f6666e9aa271d7c08dbc7e602cc