[modular][future] Fix use after free of a std::string.
The trace name of operations is meant to come from a static C string
literal. Therefore, it must not be initialized with the c_str() of
a local std::string instance, because then the const char* is accessed
after the std::string instance that backs it is destroyed and then
arbitrary memory is accessed.

Collateral cleanup:

* Move the trace_name argument to be the first argument
  of WrapFutureAsOperation(), to match the position of the trace_name
  argument everywhere else.

* Brush up some comments.

* const

TESTED=SQ

Change-Id: Iac072cf32c8f44563d631e1efb1b7acf508bef2d