[zxcrypt] Encapsulate worker requests
This CL moves the logic of building zx_port_packet_ts to be sent to
the Workers into Worker::MakeRequest.  This allows Worker to
internalize the details of these messages, and makes it easier to
add more message types in the future (as is anticipated to allow
re-keying).

SEC-26 #comment Cleanup prior to adding key mechanism

Test: zxcrypt unit tests
Change-Id: I72962354fc40b9ffafab647836e266e3bdc944a2