[audio][mixer] Add persistent ignore flags to settings.
Add flags to the audio device settings schema which allow devices to
be ignored in various fashions.  These flags are "ignore_device" and
"disallow_auto_routing".

:: ignore_device ::
When true, causes the mixer to immediately remove the device and
release the driver during activation.  Users will never see the device
show up.

:: disallow_auto_routing ::
When true, causes the mixer to never select a device when implmenting
the last-plugged automatic routing policy.  The device will still
enumerate, but it will not become the "default", and
renderers/capturers will not be automatically routed to it.

Also in this CL...
++ Allow comments in jsx files.  Comments will be removed by the
   make_schema script, along with as much whitespace as possible).
++ Improve debugging info when parsing/validating JSON save files.
   Parse failures will now give a human readable error code and an
   offset.  Schema validation errors will render a JSON object created
   by rapidjson with details about the schema validation failure.

Test: Manual on Intel NUC with manual settings file editing and server
      bouncing.
Change-Id: I7f27bd079c5a311cf50f41539f91feb1921c8ee4