[zircon] Add zircon_build_abi_dir, zircon_use_asan
This changes the zircon.gni variables around a bit so that there's one
that's for Zircon's deployment artifacts and another one that's for
compiling and linking against (sysroot and ABI shared libraries).

The new zircon_use_asan variable switches the default for the
deployment artifacts, matching the new //scripts/build-zircon.sh -A
behavior.

A new fx.config variable gives suggested switches for running
//scripts/build-zircon.sh to match the GN configuration.

Bug: SEC-27 #comment GN integration with ASan builds of Zircon
Test: Builds with coordinated //scripts/build-zircon.sh change
Change-Id: Ib249ebc60910438c5cc2b968c251edc4c61e45c3