[devmgr] Revamp pkgfs launching
The pkgfs server lives in blobfs and its dependencies (dynamic linker
et al) should come from there too.  It can reside in a proper Fuchsia
package.  Since it's what provides the filesystem layout a Fuchsia
package uses for its own executable and loader-service-provided files,
it has a bootstrapping issue.

Make fshost provide a trivial "pathname" lookup and loader service for
loading the pkgfs implementation executable.  The mapping from
pathname to blob is embedded directly in /boot/config/devmgr in the
form of `zircon.system.pkgfs.file.PATH` environment variables.
`zircon.system.pkgfs.cmd` gives the command line (in the `+` style of
the `autorun` variables) to launch the executable found this way.

`zircon.system.blob-init` and `zircon.system.blob-init-arg` are left
as a fallback when `zircon.system.pkgfs.cmd` is not supplied.  After a
transition period while the Fuchsia build is converted over to the new
method, the old support can be removed.

Bug: SEC-27 #comment pkgfs launched with correct libraries
Test: CQ + booted a Fuchsia system with GN changes that use the new method
Change-Id: Ib159fdc9e19bf757ef64aaf3ff0302b78b7be5d4