[bt][l2cap] Add dynamic channels management
Add DynamicChannel, a base class for a logical link to keep track of the
lifetime of a single dynamic channel, specifically when it becomes open
and available for user data transfer, and when it stops being open.

Add DynamicChannelRegistry, a base class for a logical link to allocate
DynamicChannel instances within the space of possible dynamic channels,
according to that link's available channel identifiers.

NET-890 #comment

Tested: Add L2CAP_DynamicChannelRegistryTest to bt-host-unittests to
test the generic behavior of allocating, opening, and closing dynamic
channels that are not specific to a kind of dynamic channel (LE or ACL).

Change-Id: I1e3be0550d695faf5f0421084907dd0170d137ea