[debugger] Add "until" command.
Exposes one-shot breakpoints in the client. Actually removes
breakpoints that the backend has marked for deletion. Adds tests for
this case.

A helper mananges the lifetime of the breakpoint, scoping it to the
lifetime of the thread or process. Currently this class manages the
lifetime of itself which is a bit weird, we will probably need to
revisit the design of this when we have more thread execution controls.

TEST=unit

Change-Id: Ic02e000b8c66aa428b2ac5ab707c8a537c076049