[callback] Allows passing lambda directly to TRACE_CALLBACK
TRACE_CALLBACK is a macro and need to use the passed parameter twice
inside its body. If the passed callback is a lambda, it would be written
twice and would have 2 different types which will prevent the ternary to
compile.

This CL solves this issue by using an anonymous lambda and storing the
passed callback in an intermediary variable.

This also allows us to remove ToStdFunction and EnsureCopyable that were
only used for this.

TEST=Everyhing still compiles

Change-Id: I3f48345e921a656c9b7b804909c43a4fa9e291c3