[board][gauss] Remove support for USB dwc3 controller
At the time the origininal code was written, I didn't realize that
the A113 SOC used dwc3 for host only and used dwc2 for device mode.
We don't support dwc2, so we only run gauss in host mode.
dwc3 host support is the same as XHCI, so we can just use the
xhci driver directly and do not need the dwc3 driver at all.

Test: manual testing on gauss. USB ethernet dongle enumerates successfully.

Change-Id: I4d90e8adb5e169aba9ac10778ad2ea8c9794973e