[dev][usb][dwc3] Make dwc3 driver responsible for starting and stopping XHCI
The dwc3 controller supports both host and device mode.
The dwc3 host support is the same as XHCI, so we use the XHCI driver
when in host mode.
Previously the board driver was responsible for starting and stopping
the XHCI driver when switching in and out of host mode.
Now the dwc3 driver does this, simplifying the support needed from the board drivers.

Test: Manual testing on hikey960 and madrone. USB host starts successfully at boot,
      "usbctl mode device" switches to device mode and "usbctl mode host" switches
      back to host.

Change-Id: Idbeab8264cd805b66b5ad6d722807e4611f7a030