[ddk][ddktl][bad-block] Add bad block protocol
This is a suplemental protocol to the nand protocol to allow for
interacting with a bad block store, which, in our case (gauss/astro), will be
a bad block table stored in a subset of the partition map. The interface
tries to make no assumptions about how the bad block information is
stored.

Keeping the interface synchronous, and keeping the MarkBlockBad limited to
a single block at a time were deliberate choices. The API should never be
in the hot path and these decisions help keep the API simpler.

It is not expected that an FTL implementation would use this, but would
be fine if it does. More typically, FTL's manage their own bad block
lists.

Change-Id: I44b5da3b13a2b02f76b7429dc9402e1910815f8d
Tested: Manually tested with future changes.