[dev][skip-block] Implement skip-block driver.
The skip-block driver attaches to non-FTL managed nand partitions exposed by the
nandpart driver. It allows for reading/writing at erase unit granularity.
This greatly simplifies the API as there is no distinction between write
and erase sizes, and therefore no need for garbage collection, or
figuring out how to append pages to the end of a partially written
block.

The API interface to skip-block is IOCTL based, and expected to be used
just by the paver utility. It should allow for updating of the
ZIRCON-A/B/R partitions, SYSCONFIG, and other non-managed partitions.

Change-Id: I4be682ae2e5c2ffa54bef27e1bcabf34bbcd795f
Tested: TODO