[cloud_provider_firestore] Make test factory take the user ID.
This patch refactors the testing factory of real Firestore cloud provider
instances exposing a way for the client to pick the user ID to use
themselves.

This makes it possible to use the factory for sync benchmarks (in an
upcoming CL), where we need multiple cloud provider instances refering
to the same cloud user.

LE-368 #comment [cloud_provider_firestore] Make test factory take the user ID.
Test: run the validation test suite against cloud_provider_firestore

Change-Id: I36d8ddec830a991c3ed56713b43b468fb2946e28