[images] Overhaul image building with ZBI
Populate /boot/config/devmgr with everything that can go there rather
than in a kernel command line (BLD-30).

Generate the BOOTFS and system-image package manifests together in one
shot so that the presence of variant drivers in the system image can
affect which devhost binaries go into the BOOTFS (ZX-2083).  Amend the
config with `devmgr.devhost.asan=true` when there are any ASan drivers
and `devhost.asan.strict=false` when there are any non-ASan drivers.

Use new `zircon.system.pkgfs.cmd` and `zircon.system.pkgfs.file.*`
arguments in /boot/config/devmgr to find pkgsvr and its libraries.

Build a single unitary complete ZBI with a single run of the `zbi`
tool.  Many kinds of booting and paving are still booting or packing a
separate kernel binary with the ZBI as the "inirtd" image.  Until
those are all converted over to the new style of boot loader support,
there will be an extra copy of the kernel in the ZBI that's never run.

Bug: BLD-30 #done
Bug: ZX-2083 #done
Bug: SEC-27 #comment GN build support for ASan drivers with non-ASan Zircon
Change-Id: Ibf0e609d55c062972c10fd7a26fcd2de02657276