[bt][l2cap] Open outbound BR/EDR dynamic channels
Add BrEdrDynamicChannel implementation of DynamicChannel that uses a ACL
signaling channel to connect and configure an outbound dynamic channel
for the corresponding link. It contains some framework to be used to
manage an inbound channel but the appropriate commands are not
implemented.

Add BrEdrDynamicChannelRegistry that creates BrEdrDynamicChannels,
dispatches inbound signaling channel request commands to the appropriate
channels, and responds to the Information Request command, which is
link-specific rather than channel-specific.

The integration into the L2CAP API and ChannelManager will come in a
subsequent CL.

NET-890 #comment

Tested: Add L2CAP_BrEdrDynamicChannelTest to bt-host-unittests with
tests for some of the branches and states possible for connecting,
configuring, and closing outbound channels.

Change-Id: Iea21935209a8089cdff42f2caa88e985b48239e3