[x86][hypervisor][vm] Avoid lgdt/lidt instructions on every VM exit
On VM exit CPU implictly sets GDT limit to 0xffff even if before
VM entry it was different. So we have to call LGDT on VM exit to
restore the limit.

However GDT limit is just a safe guard that causes GP fauilure if
someone tries to load a segment register with a value pointing
outside GDT. It's not accurate though, for example, space allocated
for GDT and the value of GDT limit depends on the value of
SMP_MAX_CPUS constant, and if SMP_MAX_CPUS doesn't match perfectly
the actual number of cpus, we end up allocating more memory than
needed.

So instead of caring about preserving GDTR limit let's just reserve
0x10000 bytes and map just enough pages to store the actual GDT and
leave the rest unmapped. One benefit of this is that LGDT is a
serializing instruction according to Intel SDM.

Note that unlike IDT we are currently modify GDT in the hypervisor
code after VM exit to restore TSS limit, so we use rw mapping around.

With IDT it's simpler, because it doesn't seem that IDT limit
affects anything unless it's smaller than 256 * 16 - 1, so we can
just get rid of lidt altogether.

Test: manually on Intel NUC and QEMU/KVM, k ut all, runtests

Change-Id: I00f87c24621e38d688859e7bbcea775a5b3f0f80