[launchpad] Fix a ZX_ERR_BAD_HANDLE: Don't return a zx_status_t as a zx_handle_t
system/ulib/process-launcher/launcher.cpp expects
launchpad_use_loader_service() to return a valid handle valid (or
ZX_HANDLE_INVALID), but it can sometimes return an error value
instead.  This happens in /boot/test/sys/spawn-test, and this causes
the test to fail if ZX_ERR_BAD_HANDLE is made a fatal error.

Fix this by not attempting to return an error in this case, since this
isn't appropriate for the interface as currently defined.

Test: /boot/test/sys/spawn-test with assertion added to
zx::object_base's close()

Change-Id: Ib8d8c9d1b638e6dbe9005596239bc7292f53a729