[fidl] Actually build a "simple" client binding
This CL marks the crash analyzer interface as simple, which triggers
code generation for the simple C client. Having this example will keep
us honest as we add the C server bindings.

To make this build, I added the make definitions need to build the C
client. This change had the result of checking that all the C headers
we've been generating actually compile. Unfortunately, the tracelink C
header did not compile, so I had to fix up the code generator to
generate more correct code.

Test: Compiles
Change-Id: Idfd9c55883aae7707b548b143e6f5867528eefde