[fidlmerge] 'get options' as function rather than method
This CL renames the root.GetOption* method to getOption* functions.
The methods are problematic, because a template needs access to root
in order to invoke them. Options can have effects that are quite far
from the root, and the single-parameter limitation of templates makes
it difficult to use root methods in many templates.

TESTED: Manual test - verified C++ code generation still works

Change-Id: I6ee7797ddd096a11b08ba390d73f84ab972a33ca