[msd-arm-mali] Prioritize atoms within a connection
Within a connection, higher-priority atoms that were submitted later may
be run before lower-priority atoms that were submitted earlier.
Lower-priority atoms may be starved, but atoms from other connections
aren't affected, in accordance with the vulkan spec.

Priority inversion shouldn't be possible because the vulkan driver only
allows 1 queue family and 2 queues within that family, so if
lower-priority queue's atoms are being starved that must be because the
higher-priority queue is running.

MA-464
TEST: vim2:go/magma-tps#L1

Change-Id: I514fafb11aca44d25f01eb48fd1b91a5fd802656