[debugger] Enhance table printing.
Adds the ability for the last column of a row to spill onto later columns. Previously
the last column would always take as much space as it needed, but it was required
that each row had the same number of columns.

The practical use of this will be to print symbol information in a disassembly
dump, where there are rows with assembly that should all be aligned horizontally,
interspersed with rows of text containing the source code information.

TEST=unit

Change-Id: I55eda639af641957d04d6c18e89681be4f1e1b7e