[ledger] Rewrite the backlog flow to better record upload time.
Rewrite the flow of the benchmark as follows:
First, the ledger is opened without sync, and gets populated
with the data. Then this connection is closed, but another is opened
with the same ledger data, but with cloud sync enabled. From that
moment, we can record the upload time. After that, the "reader" ledger
connection is opened to download the data and verify the contents.

In addition to a cleaner upload time recording, it allows to capture the
size of the ledger directory three times: without sync, with sync and
uploading, with sync and downloading.

TEST: trace record --spec-file=/pkgfs/packages/ledger_benchmarks/0/data/backlog.tspec --append-args="--server-id=<server-id>"
(on the old snapshot, while LE-508 is not resolved).

LE-488 #done

Change-Id: I696c8515e22ab3b51f1481a8a82653a1fcab8ccc