[amber] Allow amber_ctl add_src to GET its config
This change extends amber_ctl add_src to accept a URL to a JSON config
file as its "-f" parameter instead of a path to a local file.  To guard
against man-in-the middle attacks, this new option requires the caller
to provide a SHA256 hash of the config file via an "-h" parameter.

PKG-144 #done

TEST=manual

Change-Id: I750e155917928728a8f27e143721409853a432af