[pm] fix archive command consuming new manifests
The old pm manifest format was "entries separated by line terminators", the
new format is "entries in lines", which means there is a "trailing" line
terminator. The change that triggered this bug was the build change that
altered the manifest format, but ultimately the code that is changed here was
incorrect behavior.

Test: manual. This code clearly needs tests.

PKG-148 #done

Change-Id: I85c7719e5c8d78d941b3243e2d6d29a9c68abcce