[guest][cmx] Add cmx for guest_packages
We'll copy over a generic guest_package.cmx based on the guest_package()
call's target_name. This means that all guest_packages will share the
same cmx, even if their components are different. This is okay for now
because all their contents are the same. If in the future guest_package
A needs a different sandbox rule than guest_package B, then we must
break this up into A.cmx and B.cmx, checked in separately.

CP-37 #progress

TESTED=

Change-Id: Ibbb21fc336b513a615b90e21de459e60bfc7c9d3