[debugger] Improve IPC
Fixes a bug in the debug agent where only one message per read was
processed. This resulted in a strange out-of-sync effect if messages
were sent in quick succession or if the backend was slow to respond to
one.

Adds error detection and notifications to the file handle watchers. This
allows the debug_agent to close its state and wait for a new incoming
connection (previously it could accept one client connection and would
then have to be killed).

Make the file descriptor only watch for writable state when there is
data to write. I originally thought this event was edge triggered and
this would be OK, but it's actually level triggered on both POSIX and
Zircon, and this results in a storm of background "writeable"
notifications that slow things down.

TEST=manual

Change-Id: I2cc79d6997f8a1552fe8f09851ef43fb481c9645