[zx] Prepare to use typed clock for zx::time
This CL introduces zx::base_time, zx::time_utc and zx::time_thread as
alias to zx::time.

It also introduces a new templated factory method to produce
zx::base_time.

This is in preparation for the time type to be different depending on
the clock used. After this CL, clients will be able to use both the
current API, as well as the typed API. When all clients have migrated to
the new API, the old API will be removed and the time types will be
enforced.

TEST=Everything still compiles

ZX-2338

Change-Id: I3b6f4eced5df23b8dc4b1310bd8d0a89291b07a8