[ledger] report sync state as PENDING during setup
This patch changes the policy so that we report synchronization state as
PENDING during the initial setup. We also clarify the FIDL docs to state
that PENDING might mean "waiting to determine if sync is needed".

This enables clients to use synchronization state IDLE as a signal that
synchronization is completed (ensuring they don't see it prematurely).

LE-369 #done

Test: ledger_integration_tests
Change-Id: I7485f3da1d6effb05f496b7e4d49271caa57655d