[modular] Update intent structure
This updates the Intent struct to contain the action and explicit
handler directly. Previously these were conflated in the
IntentAction struct due to the module resolver using Intents for
creating queries. The module resolver has been cleaned up so now the
Intent struct no longer needs to have both be optional.

TESTED=Existing tests, as this is a mechanical change.

Change-Id: I28be3bf6acccff4f83bd9bf6cabe44a23ddaeef1