[amlogic][rawnand] Optimize delays in the Amlogic NAND controller driver.
We inject delays after writing certain ONFI commands to the controller.
This change reduces the delays. The delays are NAND chip specific (and
are different for the 2 NAND chips we use currently). For future NAND
chips, these will need to be experimentally determined.
Bug: ZX-2340

Test: Run the NAND stress tester several times on all the NAND chip
variants we plan to ship and make sure the tests pass.

Change-Id: Iff99ec8a369d83930a22ac3ff60f5dad9e07f22c