[trace-provider] Improve management of checked-in machine generated files
These files are machine generated and are checked in:
trace_provider.fidl.cpp, trace_provider.fidl.h
They are checked in to work around these problems:
- zircon's build system doesn't support transitive dependencies
- duplicate copy of tracelink is generated and compiled in garnet
- generated code for Interfaces is different than what we're using:
  struct vs zx_handle_t

Previously the machine generated contents were manually modified
post-generation. Now machine generated contents are not touched,
and we use them as-is, with the exception of the addition of comments
at the top of the files to explain things.

The next obvious step is to not check them in at all. Baby-steps.

Change-Id: I1ee748da26f5d2359c57e7b61b731907e86e169f
Tested: trace record --duration=2 /system/bin/trace --help