[kernel][arm64][boot-shim] Make the shim also a ZBI trampoline
This changes the boot shim's behavior so that if it receives an
"initrd" image from the native boot loader then it will use that
ZBI in preference to anything appended to the shim itself.

If the "initrd" is a complete ZBI with its own ARM64 kernel, then
the shim can be a bare shim (with no kernel or ZBI appended).
Anything appended to the shim is ignored in this case.

Otherwise, the kernel appended to the shim will be used but other
ZBI items appended will be ignored in favor of the "initrd" ZBI.

Bug: ZX-2224 #comment boot-shim supports trampoline mode
Test: CQ and manual booting of complete ZBI via bare shim in qemu
Change-Id: Ied0ac07840fded45b717e09f18870ab222fec126