[scripts] Give run-zircon -z and -t switches to support complete ZBI booting
The new way to boot with QEMU is to use the trampoline loader
(multiboot for x64 and qemu-boot-shim for arm64) and a complete ZBI.
The -z switch tells run-zircon the complete ZBI to boot, and the
optional -t switch tells it the trampoline binary (otherwise the
standard one is found in the build directory given with -o or
default).  The old method is still the default.

Bug: ZX-2224 #done
Bug: ZX-2225 #done
Test: CQ and manual runs with -z
Change-Id: I3e57e90dc33f3e3986cf5a61f2f2b0db34975251