[escher] Introduce FrameManager.
One of the main goals is to track the number of outstanding
frames, so that the Escher demo apps can throttle their frame
rate instead of letting the Vulkan swapchain do it.  On Linux
the NVIDIA driver allows the app to queue up 10 frames worth
of work; this interferes with a resource-recycling scheme I'm
experimenting with.

UniformBufferPool is modified to not immediately make buffers
available for reuse, instead waiting for a configurable number
of frames.  We don't yet make use of this functionality;
ModelData is configured to keep the old behavior.

Finally, fix benchmarking in Waterfall demo for shadow map mode.

ES-27 #done
ES-87 #comment
ES-100 #done

TEST=CQ, manual test of benchmarking in Waterfall demo.

Change-Id: I796939f418e13e71036bd0e23cd2e35285da4d27