[debugger] Script loading from flags
This change resulted to be quite a bit of work to get it to work
reliably. In a nutshell:

- Introduces the concept of Action. It's basically a functor that, for
now, gets generated by the processing of flags. A flag can generate
several actions (as is the case os script file). Actions have priorities
so some can run before others.
- Actions are owned by ActionFlow, which is in charge of scheduling
those actions in the MessageLoop. Each action run is given a callback to
a global function that has a reference to a global ActionFlow. That way,
when an action finishes, the ActionFlow has the context to know which
action to schedule next.
- Often, these actions cannot call the callback directly, normally
because they call verb commands. In order for this to work, verbs were
expanded to be able to receive callbacks. Not all of them can work
easily that way (Step), but others are simple to modify. Expanded the
concept of the verb action to either exist with or without callbacks. To
update one is to simply change the signature of the Do* function.
- If the console dispatches an action that doesn't receive a callback it
will call the callback instead letting the caller know that there is no
caller available. That way the ActionFlow knows that it cannot keep
running the script and falls back to interactive mode.
- When a script completes, a step has an error or a verb with no
callback was called, interactive mode will be called.

TEST=unit, manual

Change-Id: Ie87ec64f071238a3764931bee863198eea05f5c7