[zircon][aml-sd-emmc] Copy data byte by byte in the case of non dma data request
For sdio devices, we have data requests with less than 8 bytes size. This
changeset modifies the copy to be byte by byte to avoid stomping
buffers more than what is requested. This should not effect the
performance of emmc much as we generally use dma and do non-dma
transfers only for smaller sizes.

Tested: Verified exact amount of data is written in the buffers.

Change-Id: I176c9b0985794d71f116361a0ddd1336efc689ed