[appmgr] First sandbox tests
Introduce mechanism to easily exercise a sandbox feature in a test.
Introduce first tests to use this, covering the persistent-storage
feature. Many more to come.

Note that we don't actually write anything to persistent storage,
because:
1. We don't care if persistent storage actually works or is actually
persistent, just that it's there.
2. We don't want side effects from tests, especially when they affect
future invocations of the same test.

Suggestions on how to make this test case more interesting are welcome.

CP-42 #comment

TEST: {has,no}_persistent_storage

Change-Id: I2d27aba3b61af6fd6258fb0a199ec905b2c08aa2