[debugger] Listen for module loads.
Registers on process startup with the dynamic loader so the loader
issues a debug break once it's initialized the loader debug state. This
allows the debug_agent to automatically send module load notifications
to the client.

This doesn't handle shared libraries that are dynamically loaded after
starup. It will need a separate registration.

This operation currently appears as a thread stop in the middle of the
loader. This is a bnit surprising to the client. In a followup, the
client will be modified to automatically resume program execution on
this type of stop.

TEST=manual,unit

Change-Id: Ib937d2f0ad5afb972344512d45cfc6e5a4574a4d