[loader] Allow break on loader initialization.
The loader sets the process property ZX_PROP_PROCESS_DEBUG_ADDR, when it
initializes. This can be used for debuggers to get the list of loaded
libraries and to set breakpoints on library loads.

But since this is set programatically by the process code itself, the
debugger has no way to know when the query this property. If it does not
get queried before user code runs, symbolic breakpoints will not work.

This patch adds a constant that the debugger can set this property to
before execution starts. The loader checks for this and if set, issues a
manual debug breakpoint after the value is set.

A debugger can then catch this breakpoint, complete setup with the
correct debug address, and continue program execution.

TEST=manual

Change-Id: I63fc824531d1f7b09029ea01be5a93e1e9abddbb