[bt][host] RemoteDeviceCache: expire old devices
If a temporary device has not been seen in 60 seconds, expire that
device from the RemoteDeviceCache. The value of 60 seconds is chosen
somewhat arbitrarily.

We infer that a device is 'seen' by either a) the creation of the
RemoteDevice, or b) a call to a RemoteDevice method which provides
data from an inquiry. The lifetime is extended even if the new data
is invalid.

The device deletion tasks are tracked within RemoteDeviceCache.
Alternatively, we could make the deletion task part of RemoteDevice.
The choice to track tasks in RemoteDeviceCache is due to a desire to
limit reentrancy during destruction. Said differently, we would like
to avoid the following:

   void RemoteDevice::ExpirationClosure() {
     void RemoteDeviceCache::ExpireDevice(...) {
       void RemoteDevice::~RemoteDevice() {
         // SUBTLE! May be called while a RemoteDevice method
	 // is on the stack.
       }
     }
     // SUBTLE! May execute after destructor
   }

The approach taken here doesn't eliminate the problem (since there
is stil subtle code within RemoteDeviceCache), but it does isolate
the problem to just one file.

Note that this CL also changes the existing GAP_RemoteDeviceCacheTest
suite of tests to use a test fixture. This is necessary so that
async_get_default() returns non-null when we UpdateExpiry() as
part of NewDevice().

While there:
- Remove RemoteDeviceMap typedef, as it was only referenced in one
  place.
- Make comments for RemoteDeviceCache::devices_ and
  RemoteDeviceCache::address_map_ symmetric.  ("Mapping" in both
  places.)

Test: bt-host-unittests (on QEMU)
Test: added GAP_RemoteDeviceCacheTest_ExpirationTest.*
Test: manual
NET-893 # done

Manual testing
--------------
Ran the following shell commands, and didn't see any "FAIL" messages.

$ discovery_time=60
$ expiry_time=60
$ total_time=$((discovery_time + expiry_time + 1))
$ sleep $expiry_time
$ fx shell '
    (echo start-discovery; sleep 5; echo stop-discovery;
     for i in $(seq 0 '$total_time'); do echo adapter; sleep 1; done) \
    | bt-cli 2>&1' \
  | awk 'BEGIN {start=systime()}
        /\{/ {event=$0; next}
        {print systime()-start, event, $0; event=""; fflush()}' \
  | tee /tmp/discovery.log
$ n_found=$(
    awk '/RemoteDevice/ { print $6 }' /tmp/discovery.log | sort | uniq | wc -l)
$ n_unique_removals=$(
    awk '/Device .+ removed/ { print $3 }' /tmp/discovery.log | sort | uniq | wc -l)
$ n_total_removals=$(egrep 'Device .+ removed' /tmp/discovery.log | wc -l)
$ first_removal_time=$(
    egrep 'Device .+ removed' /tmp/discovery.log | head -1 | cut -d ' ' -f1)
$ (( n_found == n_unique_removals )) ||
    echo "FAIL: Found $n_found, but removed $n_unique_removals"
$ (( n_unique_removals == n_total_removals )) ||
    echo "FAIL: Duplicate removals: $((n_total_removals - $n_unique_removals))"
$ (( first_removal_time >= 60 )) || \
    echo "FAIL: Early removal at $first_removal_time"

Change-Id: Ibb26d672c32dda654e858b5b87a62cb955456c6b