[debugger] Fix some crashes.
The first crash is an assertion failure checking for self-assignment of
a watch handle. The code was doing an assignment of a null handle when
handling a connection error which should be allowed. The assertion has
been updated.

This was hit for the first time due to a crash in the debug agent
getting the address space. The vector resize call got lost when the code
was moved to a separate function.

TEST=manual

Change-Id: If5f142c3f2941613ec072a5e1ef59bb54273ef5a