[debugger] Add backtrace and list commands.
Added a "verbose" mode to the "frame" command which lists frames.
Currently this just adds the instruction pointer, but in the future I'm
planning on using it to add more stuff like function parameters.

A "backtrace" command is added which is an alias for "frame -v". "bt" is
what people are used to typing in GDB, so it's important that this work.

Noun handling is somewhat refactored. The help and registration has been
moved from the general command.cc to nouns.cc where it makes more sense.

Support is added for passing switches to noun commands.

Frame list printing is enhanced somewhat. Since it can be hard to find
the frame number in a long backtrace with a lot of templates in the
function names, the frame numbers are now printed in reverse video.

A "list" command is added to list the source code at the current
location, or at a given one. The usage should be very similar to GDB.

The source code listing function now separates the "active" line from
the "highlighted" line since I wanted to highlight the symbol the user
requested, but reserve the ">" in the left column for the current code
location. This highlighting is now used for breakpoints also which looks
a bit better.

TEST=manual,unit

Change-Id: I5fd98425569b925f84ee6cb750017a27ceb40d83