[chrealm] Better simulation of realm environment.
Currently, we only replace the /svc directory in the chrealmed process.
This CL contains two changes that make the chrealmed process behave
more like an actual subprocess of the realm:

- Replace /hub with the hub for the realm.
- Spawn the process under the realm's job. This makes it so that if the
realm is killed, the process dies along with it.

To support #2, we've added a "job" service under the hub that supplies
a handle to the realm's job.

CP-54 #done

TESTED=chrealm_integration_test

Change-Id: Icef6dd221b54fded7e2f4e7d4414bb62024bc787