[vim][audio] Give audio HW resources to HDMI
Give the HW resources needed to control the audio units over to the
HDMI display driver.

Right now, there is no really good way to separate the audio driver
from the HDMI driver and still have HDMI expose audio functionality
(we need some support for composite drivers in the devmgr/ddk to
actually do this properly), so for now, we just give all of the audio
resources over to the display driver.

Test: Build and manual integration with WiP VIM2 audio driver.
Change-Id: Ie0320302cd84a3a9c917ed71ab672c7d9b1dda4f