[vim2][display] Fix mmio table alignment
Fix the MMIO enumerations in the vim-display driver so that they line
up with the MMIO pdev table being passed by the board driver.  Also,
add a runtime check to verify that the number of entries in the
enumeration at least matches the number of regions passed by the board
driver.

In general, it seems like it might be good if we could come up with a
less brittle way of passing and posititvely identifying these
resources, but I suppose that is a challenge for another day.

Test: Built and booted on a VIM2, verified functioning HDMI output.
Change-Id: I6bc248cf3ee9c00c7e8b2c44267f1a4ff22b3fa2